requirements.txt
certifi==2020.12.5
chardet==4.0.0
dnspython==2.1.0
idna==2.10
python-pushover==0.4
requests==2.25.1
urllib3==1.26.3
pytest==6.2.2